refactor: 카드컴포넌트 모듈
