+++ client/src/components/ui/cards/MapLocationCard.tsx
@@ -122,7 +122,7 @@ const LocationA = styled.a`
   }
 `;
 
-const MapLocationCard = ({
+export const MapLocationCard = ({
   indexNum,
   location,
   latlng,
@@ -186,5 +186,3 @@ const MapLocationCard = ({
     </MapLocationCardContainer>
   );
 };
-
-export default MapLocationCard;
